% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFile.R
\name{GetFile}
\alias{GetFile}
\title{Loads a data frame using the OS File Dialog}
\usage{
GetFile()
}
\value{
Depending on the choice selected the function will return: \cr
\cr
1) The path of the file; \cr
2) An R data frame; \cr
3) A data.table data frame; \cr
4) An Arrow data frame; \cr
}
\description{
Lets you pick a file in your computer using the File Dialog box and loads it as a data frame, data table, or arrow frame.
}
\examples{
\dontrun{
MyFilePath <- GetFile() #If file path radio button selected.
MyDataFrame <- GetFile() #If one of the Data Frame options selected in the dialog box.
}

}
\author{
Rodrigo Pinto Coelho
}
